Move response handling steps
